+++ PeopleDieGame.ServerPlugin/Commands/Admin/ManageShopCommand.cs
@@ -88,8 +88,8 @@ namespace PeopleDieGame.ServerPlugin.Commands.Admin
                 ShopManager shopManager = ServiceLocator.Instance.LocateService<ShopManager>();
                 ShopItem shopItem = shopManager.ResolveItem(command[0], false);
 
-                double price;
-                if (!double.TryParse(command[1], out price))
+                float price;
+                if (!float.TryParse(command[1], out price))
                 {
                     ChatHelper.Say(caller, "Artykuł 13 paragraf 7 - kto defekuje się do paczkomatu");
                     return;
@@ -180,8 +180,8 @@ namespace PeopleDieGame.ServerPlugin.Commands.Admin
                 return;
             }
 
-            double price;
-            if (!double.TryParse(command[1], out price))
+            float price;
+            if (!float.TryParse(command[1], out price))
             {
                 ChatHelper.Say(caller, "Artykuł 13 paragraf 7 - kto defekuje się do paczkomatu");
                 return;
